Simple change that allows PASSIVE_SANITY_SOURCES to be modified.
With the previous way the array "PASSIVE_SANITY_SOURCES" was made, you could not add or remove values from it because it was a fixed size list. This change turns it into a real ArrayList allowing modders to use methods like add() or addAll() on PASSIVE_SANITY_SOURCES in order to add their own sanity sources.